+++ .pre-commit-config.yaml
@@ -31,5 +31,5 @@ repos:
       hooks:
           - id: black
             language_version: python3.6
-            args: [-l=100, --exclude='/(\..*)/']
+            args: [-l 100]
             files: 'backend/tracim_backend'
