+++ backend/setup.cfg
@@ -24,7 +24,7 @@ force_grid_wrap = 0
 use_parentheses = true
 skip_glob = .eggs*,.venv
 known_standard_library = contextvars,dataclasses
-known_third_party = PIL,alembic,babel,bs4,caldav,cliff,colour,depot,email_reply_parser,filelock,freezegun,hapic,imapclient,ldap3,lxml,mako,markdown,marshmallow,mock,parameterized,paste,plaster,plaster_pastedeploy,preview_generator,pyramid,pyramid_ldap3,pyramid_multiauth,pytest,pytz,radicale,redis,requests,responses,rq,setuptools,slugify,sqlalchemy,transaction,waitress,webtest,wsgi,wsgidav,yaml,zope
+known_third_party = PIL,alembic,babel,bs4,caldav,cliff,colour,depot,email_reply_parser,filelock,freezegun,hapic,imapclient,ldap3,lxml,mako,markdown,marshmallow,mock,parameterized,paste,plaster,plaster_pastedeploy,preview_generator,pyramid,pyramid_ldap3,pyramid_multiauth,pytest,pytz,radicale,redis,requests,responses,rq,setuptools,slugify,sqlalchemy,transaction,waitress,webtest,wsgidav,yaml,zope
 known_first_party = tracim_backend
 
 [flake8]
