+++ backend/wsgi/webdav.py
@@ -1,7 +1,7 @@
 # coding=utf-8
 # Runner for uwsgi
 import os
-from wsgi import webdav_app
+from tracim_backend.wsgi import webdav_app
 
 config_uri = os.environ['TRACIM_CONF_PATH']
 application = webdav_app(config_uri)
